Revert _rand_buf type reversal

+++ pxtone/pxtnPulse_NoiseBuilder.h
@@ -13,7 +13,10 @@ private:
 
 	bool    _b_init;
 	short*  _p_tables[ pxWAVETYPE_num ];
-	int32_t _rand_buf [ 2 ];
+	// OPNA2608 EDIT
+	// see pxtnPulse_NoiseBuilder.cpp comment for explanation
+	// int32_t _rand_buf [ 2 ];
+	uint16_t _rand_buf [ 2 ];
 
 	void  _random_reset();
 	short _random_get  ();
